requirements.txt
Django==1.4
boto==2.5.2
cssmin==0.1.4
django-pipeline==1.2.10
django-storages==1.1.4
jsmin==2.0.2
